update build script

+++ build.sh
@@ -7,13 +7,17 @@ then
 
 	./dist/gopy
 
-	echo "--done--"
-else
+	echo "done..."
+elif [[ $1 = "build" ]]
+then
 	echo "building..."
 	
 	rm -r ./dist
 	cd src
 	go build -o ../dist/gopy
 
-	echo "--done--"
+	echo "done..."
+else
+	./build.sh build
+	./build.sh run
 fi
